extracted PrivateRoute and AuthRoute to separate files
