+++ src/components/Main.js
@@ -1,9 +1,12 @@
 import React from 'react';
-import { Switch, Route, Redirect } from 'react-router-dom';
+import { Switch } from 'react-router-dom';
 
 import Dashboard from './Dashboard/Dashboard';
 import BookDetail from './BookDetail/BookDetail';
 import Login from './Login/Login';
+import PrivateRoute from './routes/PrivateRoute';
+import AuthRoute from './routes/AuthRoute';
+
 import './Main.css';
 
 function Main() {
@@ -17,35 +20,5 @@ function Main() {
     </main>
   );
 }
-
-const PrivateRoute = ({
-  component: Component, // eslint-disable-line react/prop-types
-  ...rest
-}) => (
-  <Route
-    {...rest}
-    render={props =>
-      localStorage.token ? (
-        <Component {...props} />
-      ) : (
-        <Redirect to={{ pathname: '/login', state: { from: props.location } }} /> // eslint-disable-line react/prop-types
-      )}
-  />
-);
-
-const AuthRoute = ({
-  component: Component, // eslint-disable-line react/prop-types
-  ...rest
-}) => (
-  <Route
-    {...rest}
-    render={props =>
-      localStorage.token ? (
-        <Redirect to={{ pathname: '/', state: { from: props.location } }} /> // eslint-disable-line react/prop-types
-      ) : (
-        <Component {...props} />
-      )}
-  />
-);
 
 export default Main;
